+++ Components/MeshLodGenerator/include/OgreLodCollapser.h
@@ -69,7 +69,7 @@ namespace Ogre
         {
             unsigned int srcID;
             unsigned int dstID;
-            unsigned short submeshID;
+            unsigned int submeshID;
         };
 
         typedef vector<CollapsedEdge>::type CollapsedEdges;
@@ -85,9 +85,9 @@ namespace Ogre
         void assertOutdatedCollapseCost(LodData* data, LodCollapseCost* cost, LodData::VertexI vertexi);
         void assertValidMesh(LodData* data);
         void assertValidVertex(LodData* data, LodData::VertexI vi);
-        bool hasSrcID(unsigned int srcID, unsigned short submeshID);
+        bool hasSrcID(unsigned int srcID, unsigned submeshID);
         void removeTriangleFromEdges(LodData* data, LodData::Triangle* triangle, LodData::VertexI skipi);
-        size_t findDstID(unsigned int srcID, unsigned short submeshID);
+        size_t findDstID(unsigned int srcID, unsigned submeshID);
         void replaceVertexID(LodData* data, LodData::Triangle* triangle, unsigned int oldID, unsigned int newID, LodData::Vertex* dst);
     };
