+++ Components/MeshLodGenerator/include/OgreLodData.h
@@ -130,7 +130,7 @@ namespace Ogre
             VertexI vertexi[3];
             Vector3 normal;
             bool isRemoved;
-            unsigned short submeshID; /// ID of the submesh. Usable with mMesh.getSubMesh() function.
+            unsigned int submeshID; /// ID of the submesh. Usable with mMesh.getSubMesh() function.
             unsigned int vertexID[3]; /// Vertex ID in the buffer associated with the submeshID.
 
             void computeNormal(const VertexList& vertexList);
